requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async & HTTP
httpx==0.26.0
aiofiles==23.2.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# NLP & ML
sentence-transformers==2.3.1
scikit-learn==1.4.0
nltk==3.8.1
spacy==3.7.2

# Vector Store & Search
faiss-cpu==1.7.4
chromadb==0.4.22

# Database
duckdb==0.9.2
sqlalchemy==2.0.25

# LLM & Embeddings
openai==1.10.0
tiktoken==0.5.2
langchain==0.1.4
langchain-community==0.0.16

# Text Processing
regex==2023.12.25
unidecode==1.3.8

# UI
streamlit==1.30.0
plotly==5.18.0
altair==5.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
rich==13.7.0

# Observability & Monitoring
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.0.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
